+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Deleting images and optionally its related snapshots `awless delete image delete-snapshots=true`
 - Autowrapping results in tables when too long for `awless list`. No longer truncate results in `--format csv/tsv/json`
 - Adjust the width of table columns to the terminal width in `awless show`
+- installing awless on an EC2 instance now set the correct region using EC2 local metadata
 
 ## v0.0.23 [2017-05-05]